refactor: move nx2 RUM logic into scripts/lazy.js
Extract rumWC and the rum bootstrap out of nx.js into a self-invoking
scripts/lazy.js (mirroring nx1). nx.js now lazy-imports lazy.js after the
section loop instead of holding the logic inline.
