+++ nx2/scripts/nx.js
@@ -275,29 +275,6 @@ async function decoratePlaceholders(area, isDoc) {
   }
 }
 
-function rumWC(sampleRUM) {
-  const wcs = document.querySelectorAll('[data-rum]');
-  wcs.forEach((wc) => {
-    wc.shadowRoot.addEventListener('click', (e) => {
-      e.stopPropagation();
-      const sourceEl = e.target.closest('a, button');
-      const source = sourceEl?.title || sourceEl?.href || sourceEl?.dataset.action;
-      if (!sampleRUM.targetselector) return;
-      const target = sampleRUM.targetselector(e.target);
-      sampleRUM('click', { source, target });
-    });
-  });
-}
-
-function loadRum() {
-  import('../deps/rum.js').then(({ sampleRUM }) => {
-    sampleRUM();
-    window.setTimeout(() => {
-      rumWC(sampleRUM);
-    }, 3000);
-  });
-}
-
 function loadSession() {
   sessionStorage.setItem('session', true);
   document.body.classList.add('session');
@@ -356,7 +333,7 @@ export async function loadArea({ area } = { area: document }) {
     delete section.dataset.status;
     if (isDoc && idx === 0 && loadHeader(isSession)) break;
   }
-  if (isDoc) loadRum();
+  if (isDoc) import('./lazy.js');
 }
 
 const cache = {};
